Authenticate and authorised user with generated token to delete posts

+++ controllers/api/v1/posts_api.js
@@ -26,16 +26,18 @@ module.exports.destroy = async function (req, res) {
 
     try {
         let post = await Post.findById(req.params.id);
-
-        await Comment.deleteMany({ _id: post.comments });
-        post.remove();
-        return res.status(200).json({
-            message:"Post and its comments deleted",
-        })
-        // }else{
-        //     console.log('not authorised');
-        //     return res.redirect('back');
-        // }
+        if(req.user.id == post.user){
+            await Comment.deleteMany({ _id: post.comments });
+            post.remove();
+            return res.status(200).json({
+                message:"Post and its comments deleted",
+            })
+        }else{
+            console.log('not authorised');
+            return res.status(401).json({
+                message:"Not authorised to delete the post"
+            })
+        }
     } catch (err) {
         console.log("Error --->  ", err);
         return res.status(500).json({
